initial layout of client-profile-map component. Will fine tune layout and start fine tuning features
