+++ outreach-mapping-solutions-FE/src/app/portals/client-portal/client-profile/client-profile-map/client-profile-map.component.html
@@ -1,6 +1,23 @@
 
+
+<div class="row g-1">
+    <div class="col">
+        <app-client-map></app-client-map>    </div>
+    <div class="col-3">
+        <app-interaction-list></app-interaction-list>
+    </div>
+</div>
+<div class="row">
+    <div class="col">
+        <app-interaction-form></app-interaction-form>
+    </div>
+</div>
+
+
+
+
 <!-- Add this HTML structure for the popup and form in your template -->
-<div id="map-container" style="width: 100%; height: 400px;"></div>
+<!-- <div id="map-container" style="width: 100%; height: 400px;"></div>
 
 <div id="popup" class="ol-popup">
   <a href="#" id="popup-closer" class="ol-popup-closer"></a>
@@ -12,7 +29,7 @@
       <button type="button" id="cancel-button">Cancel</button>
     </form>
   </div>
-</div>
+</div> -->
 
 
     <!-- <div id="tools">
